requirements.txt
pandas==1.1.3
numpy==1.19.2
scikit-learn==0.23.2
grpcio==1.33.2
grpcio-tools==1.16.1